Rename 'overlay2' to 'overlay'
Finish the overlay->overlay2 changeover by:
* renaming overlay2 to overlay
  * teaching the driver to accept the overlay.override_kernel_check and
    overlay2.override_kernel_check options as having the same meaning
* registering the driver init function using both names, returning as
  the driver's name
